raidboss: refactor TimelineUI to not include HTML elements
Refactor TimelineUI into HTMLTimelineUI and base class
TimelineUI.  This puts HTMLTimelineUI in its own file
so it's possible to run test_timeline without including dom.
